Added file exits check to the post-data.py

+++ scripts/post-data.py
@@ -2,14 +2,15 @@ import urllib
 import urllib2
 import socket
 
-with open('/home/pi/probereqs.log') as f:
-    probedata = f.read()
+if os.path.exists('/home/pi/probereqs.log'):
+    with open('/home/pi/probereqs.log') as f:
+        probedata = f.read()
 
-url = 'https://bb.w4p.us/brother.php'
-values = { 'device': socket.gethostname(), 'data': probedata }
-data = urllib.urlencode(values)
+    url = 'https://bb.w4p.us/brother.php'
+    values = { 'device': socket.gethostname(), 'data': probedata } 
+    data = urllib.urlencode(values)
 
-req = urllib2.Request(url, data)
-response = urllib2.urlopen(req)
+    req = urllib2.Request(url, data)
+    response = urllib2.urlopen(req)
 
-with open('/home/pi/probereqs.log', 'w'): pass
+    with open('/home/pi/probereqs.log', 'w'): pass
